Revert "Merge branch 'expansion+ow-ex+dyn-pals' into 3-friends"
This reverts commit 013108f2f7fd22951d3687ca3fb32836ec05e979, reversing
changes made to 180b86a60f4928d6d385ca5d1c9f76a80e355b44.

+++ src/cable_car.c
@@ -797,7 +797,7 @@ static void CreateCableCarSprites(void)
         [SPENCER] = OBJ_EVENT_GFX_SPENCER_NORMAL
     };
     u16 rval = Random();
-    u16 hikerGraphicsIds[4] = {
+    u8 hikerGraphicsIds[4] = {
         OBJ_EVENT_GFX_HIKER,
         OBJ_EVENT_GFX_CAMPER,
         OBJ_EVENT_GFX_PICNICKER,
